refactor: добавит директорию utils. правит пути к файлам
