+++ project/src/components/private-outlet/private-outlet.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { AppRoute, AuthorizationStatus } from '../../utils/const';
 
 interface PrivateOutletProp {
   authorizationStatus: AuthorizationStatus;
