+++ project/src/utils/const.ts
@@ -45,18 +45,6 @@ export enum AuthorizationStatus {
   UNKNOWN = 'UNKNOWN',
 }
 
-const MAX_RATING = 5;
-
-export const getRatingPercent = (rating: number) => (rating / MAX_RATING) * 100;
-
-export const getFormatDate = (date: string) => {
-  const currentDate = new Date(date);
-  const getMonth = currentDate.toLocaleString('en', { month: 'long' });
-  const getYear = currentDate.getFullYear();
-
-  return `${getMonth} ${getYear}`;
-};
-
 export enum SortTypes {
   POPULAR = 'Popular',
   PRICE_LOW = 'Price: low to high',
